Provisional release notes support.
Add "notes" macro in openssl.wml to translate the appropriate section
in the NEWS file from an OpenSSL branch into a set of release notes.

A .wml file the form openssl-X.Y.Z-notes.wml automatically uses the
correct section from the NEWS file from the appopriate branch and
includes a generic introduction.

Defaults can also be overridden by including "minversion", "maxversion"
and "dirname" options. The introduction for this version is supplied
in the .wml page.