It seems that pod2html in more modern perl releases don't do title.
The pod2html that comes with perl 5.14 will extract the title from the
NAME section of the .pod file.  In perl 5.18, this is no longer the
case, and pod2html has a new argument --title where this can be given
manually.

The solution here is to test the behavior of pod2html, and if it
doesn't produce a title automatically, we extract it ourselves (simple
sed does it) and provide the resulting string to pod2html.