Reduce relupd to what it's actually needed for
With the help of git hooks, the checkout cache and web directory tree
get updated (*) automagically whenever something relevant is pushed.
All relupd is still needed for is to make sure the release manpages
are being updated on release (we should refrain from doing so at any
time something is pushed, because we do want the release manpages to
correspond to the latest releases rather than whatever changes that's
come later on).

(*) in the web directory tree, "make" is also run automatically.