inc/screen.css: no pre-wrap
There's no reason why the contents of <a> element should be wrapped on
line breaks.  Set white-space to 'normal' instead.

This property is useful in case we happen to inherit some other
setting of that property and want to get back to a normal setting.

Reviewed-by: Matt Caswell <matt@openssl.org>
(Merged from https://github.com/openssl/openssl/pull/79)