bin/mk-manpages: refactor to allow cross references between releases
So far, we created one HTML file for each POD file, and then made hard
links to it for other names that are in the POD file's NAMES section.

However, this came with the assumption that cross referencing between
releases would work simply be linking to the same name on other
releases.  This, however, did not take into account that manuals in
newer releases don't necessarily exist in older releases, or that some
files may have changed names.

Names in NAMES sections are, however, fairly constant, and are
therefore much safer to link to.  At the same time, it's safe to say
that if a particular name doesn't exist in some other releases, there
should simply not be a link.  A conclusion to draw from is that cross
referencing must be made on a per NAMES section name basis, rather
than on POD file name basis.

To allow this to happen and still not have to rewrite the same
Pod2Html result for every name in a specific POD file's NAMES section,
the structure of the rendered man pages are changed to this:

-   POD files are rendered into a .inc file with the exact same
    basename as the POD file.
-   For every name in the NAMES section, am HTML file is created.  It
    contains the standard header and footer stuff, and includes the
    generated .inc file in the middle.  It also includes a .cross file
    with the same basename as the HTML file as part of the sidebar.

In another commit, there will be a script for cross referencing, which
will generate the .cross files mentioned above.

Reviewed-by: Matt Caswell <matt@openssl.org>
(Merged from https://github.com/openssl/web/pull/107)