Rework the markdown to HTML script and template
The pandoc template inherits all the HTML that was included with
server side includes.  For things that depend on the directory, we use
metadata variables, which are provided externally, or as part of the
markdown origin.

bin/md-to-html5 is changed accordingly.  Among other things, it checks
for dirdata.yaml in the directory of the markdown origin, and includes
it into the build as well.

dirdata.yaml is expected to contain metadata that applies to all
generated HTML files in the containing directory.  Two metadata
attributes are expected:

- `breadcrumbs`, to give an indication of where the user is, for every
  page.  If the HTML to be produced isn't the index file, it will
  automatically be added to the breadcrumbs by the pandoc template.
- `sidebar`, which will be the content of the sidebar

Both of these attributes take strings that are interpreted as markdown