Fix bin/mk-manpages3 to handle spurious & in the description
We have some pages that emit &lt; and &gt; in the NAMES description in
the HTML output.

However, we're using sed to massage a template with that description,
and & happens to be significant.  Therefore, it needs being explicitly
escaped.

Partially fixes openssl/openssl#13949

Reviewed-by: Paul Dale <pauli@openssl.org>
Reviewed-by: Matt Caswell <matt@openssl.org>
(Merged from https://github.com/openssl/web/pull/214)