Simplify the CDN purge
Our CDN (Akamai) purge script was run as a standalone automation, in
parallell with the automatic 'make' run.  The consequence was that the
CDN could catch a copy of our original web pages in a semi built state,
as demonstrated by openssl/openssl#16152.

The solution is the ensure that the purge is run in sequence after
everything is built.  We simplify this further by moving the actual
script into the web source.

Reviewed-by: Paul Dale <pauli@openssl.org>
(Merged from https://github.com/openssl/web/pull/248)