Make an exception for the glossary
The glossary is in $(CHECKOUTS)/general-policies/policies/, but is
common for both the general and the technical policies, so we place
it directly in `policies/`.

Since policies/glossary.html is generated from markdown, the breadcrumbs and
sidebar must be duplicated in an added policies/dirdata.yaml.

Fixes #306

Reviewed-by: Tim Hudson <tjh@openssl.org>
Reviewed-by: Matt Caswell <matt@openssl.org>
Reviewed-by: Tomas Mraz <tomas@openssl.org>
Reviewed-by: Paul Dale <pauli@openssl.org>
(Merged from https://github.com/openssl/web/pull/312)