Add the necessary files and procedures to generate HTML from policies
The checked out directories $(CHECKOUTS)/technical-policies/policies and
$(CHECKOUTS)/general-policies/policies contain files in markdown format,
which get converted to HTML using pandoc with the new template, and
appropriate dirdata.yaml.

The base infrastructure in policies/general and policies/technical is also
created, and the majority of the links in policies/sidebar.shtml are moved
to policies/general/sidebar.shtml and policies/general/dirdata.yaml.

Fixes #306