Extend the primary platforms list
Add linux-x86, linux-generic32 and linux-generic64 as primary
platforms. For linux-generic64 I am proposing this to be on
Ubuntu 20.04.3 for consistency with the linux-x86_64 target. For
the 32 bit builds I am proposing Debian 11.2 since Ubuntu does not
appear to have a 32 bit version.

This implements the vote in openssl/general-policies#12

Reviewed-by: Richard Levitte <levitte@openssl.org>
(Merged from https://github.com/openssl/web/pull/314)