Better dependencies for md->html rendering
It's too easy to miss adding proper depdencies, so let's make that
automatic, at least for md files we have in this repository.

All md->html rendering is made to depend on the presence of dirdata.yaml,
which will be a good reminder to add it where it's missing.

Reviewed-by: Matt Caswell <matt@openssl.org>
(Merged from https://github.com/openssl/web/pull/326)