Don't use $? when only the first prerequisite is to be used
There are a few rules where we were using $?.  This was safe as long
as there was only one prerequisite, even though unnecessary.  However,
when there are multiple prerequisites (such as the recently added
Makefile prerequisites), we lose control on what's actually used in
the commands of the rule.

We take back control by replacing $? with $<, which is always the
first prerequisite.

Reviewed-by: Mark J. Cox <mark@awe.com>
Reviewed-by: Matt Caswell <matt@openssl.org>
(Merged from https://github.com/openssl/web/pull/332)