policies/omc-bylaws.md: Add a clause on minor edits
As per discussion on the OMC meeting 2022-07-13, a clause for minor edits
is added, to allow edits that do not change the meaning of the bylaws, such
as a conversion to markdown.

An OMC vote has been held regarding this commit, and passed.

Reviewed-by: Tim Hudson <tjh@openssl.org>
(Merged from https://github.com/openssl/web/pull/347)