support/acks.md: Remove "current" link
Originally (when this file was still raw HTML), this was an ID
(<h3 id="current">Sponsorship Donations</h3>).  Turning it into a link
was a clear typo.

As far as I can tell, nothing linked to acks.html#current, so we can
as well drop it entirely.

Fixes #371

Reviewed-by: Paul Dale <pauli@openssl.org>
(Merged from https://github.com/openssl/web/pull/372)