Update pink nirvana and related documentation.
I have seen multiple instances of references to the design document,
and in a few cases copying of the pink nirvana diagram - so it was time
to update it to match what we currently have.
Also changed text related to the engine provider, since that was
completely misleading.

Reviewed-by: Matt Caswell <matt@openssl.org>
Reviewed-by: Richard Levitte <levitte@openssl.org>
(Merged from https://github.com/openssl/web/pull/380)