Fix policy redirects
We have recently removed two policies in favor of new ones in the general/
subdirectory.  For the benefit of anyone having links to the old policy
documents, we habitually add redirects so they get to the new version.
We forgot that for the platform and travel policies...

Reviewed-by: Tomas Mraz <tomas@openssl.org>
(Merged from https://github.com/openssl/web/pull/388)