Move release table higher up
Currently the table of the latest releases for each major version is quite far down the page, even below the bottom of the viewport.

This confused me (and probably also others) as I thought the table was removed.

Move the table near the top of the page, as I feel that it's likely the most important part of the page in general.

Reviewed-by: Matthias St. Pierre <Matthias.St.Pierre@ncp-e.com>
Reviewed-by: Paul Dale <pauli@openssl.org>
Reviewed-by: Tomas Mraz <tomas@openssl.org>
(Merged from https://github.com/openssl/web/pull/414)