Fix the HTML rendering of man pages to avoid markdown reinterpretation
Some of our manpages have the construct "[something](somethingelse)".

From a POD and HTML perspective, there's nothing special with that.
However, since we wrap the HTML with markdown (perfectly legitimate), we
need to pay attention that nothing in the HTML can reasonably be
reinterpreted as markdown.

Unfortunately, the construct mentioned about has that issue.

This is easily fixed by replacing '](' with '&rbrack;(', which is valid
HTML5.

Reviewed-by: Matt Caswell <matt@openssl.org>
(Merged from https://github.com/openssl/web/pull/443)